feat: ai input generator validation
